requirements_dbwriter.txt
kafka-python
SQLAlchemy
psycopg2
